microservices/ingestion/requirements.txt
# Ingestion microservice – Python dependencies
# PySpark 3.x
pyspark>=3.4,<4.0
pyyaml>=6.0
